+++ neva/src/types/tool.rs
@@ -126,15 +126,10 @@ pub struct Tool {
     #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
     pub meta: Option<Value>,
 
-    /// A list of roles that are allowed to invoke the tool
+    /// What a caller must hold to invoke the tool.
     #[serde(skip)]
     #[cfg(feature = "http-server")]
-    pub(crate) roles: Option<Vec<String>>,
-
-    /// A list of permissions that are allowed to invoke the tool
-    #[serde(skip)]
-    #[cfg(feature = "http-server")]
-    pub(crate) permissions: Option<Vec<String>>,
+    pub(crate) required: crate::transport::http::core::auth::RequiredClaims,
 
     /// A tool call handler
     #[serde(skip)]
@@ -1296,9 +1291,7 @@ impl Tool {
             custom_schema: false,
             icons: None,
             #[cfg(feature = "http-server")]
-            roles: None,
-            #[cfg(feature = "http-server")]
-            permissions: None,
+            required: Default::default(),
             #[cfg(feature = "tasks")]
             exec: None,
         }
@@ -1416,7 +1409,7 @@ impl Tool {
         T: IntoIterator<Item = I>,
         I: Into<String>,
     {
-        self.roles = Some(roles.into_iter().map(Into::into).collect());
+        self.required.set_roles(roles);
         self
     }
 
@@ -1427,7 +1420,7 @@ impl Tool {
         T: IntoIterator<Item = I>,
         I: Into<String>,
     {
-        self.permissions = Some(permissions.into_iter().map(Into::into).collect());
+        self.required.set_permissions(permissions);
         self
     }